+++ pai-dlc-20201203/src/Models/GetJobResponseBody/userVpc.php
@@ -1,0 +1,103 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Paidlc\V20201203\Models\GetJobResponseBody;
+
+use AlibabaCloud\Tea\Model;
+
+class userVpc extends Model
+{
+    /**
+     * @var string
+     */
+    public $defaultRoute;
+
+    /**
+     * @var string[]
+     */
+    public $extendedCidrs;
+
+    /**
+     * @example sg-abcdef****
+     *
+     * @var string
+     */
+    public $securityGroupId;
+
+    /**
+     * @example vs-abcdef****
+     *
+     * @var string
+     */
+    public $switchId;
+
+    /**
+     * @example vpc-abcdef****
+     *
+     * @var string
+     */
+    public $vpcId;
+    protected $_name = [
+        'defaultRoute'    => 'DefaultRoute',
+        'extendedCidrs'   => 'ExtendedCidrs',
+        'securityGroupId' => 'SecurityGroupId',
+        'switchId'        => 'SwitchId',
+        'vpcId'           => 'VpcId',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->defaultRoute) {
+            $res['DefaultRoute'] = $this->defaultRoute;
+        }
+        if (null !== $this->extendedCidrs) {
+            $res['ExtendedCidrs'] = $this->extendedCidrs;
+        }
+        if (null !== $this->securityGroupId) {
+            $res['SecurityGroupId'] = $this->securityGroupId;
+        }
+        if (null !== $this->switchId) {
+            $res['SwitchId'] = $this->switchId;
+        }
+        if (null !== $this->vpcId) {
+            $res['VpcId'] = $this->vpcId;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return userVpc
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['DefaultRoute'])) {
+            $model->defaultRoute = $map['DefaultRoute'];
+        }
+        if (isset($map['ExtendedCidrs'])) {
+            if (!empty($map['ExtendedCidrs'])) {
+                $model->extendedCidrs = $map['ExtendedCidrs'];
+            }
+        }
+        if (isset($map['SecurityGroupId'])) {
+            $model->securityGroupId = $map['SecurityGroupId'];
+        }
+        if (isset($map['SwitchId'])) {
+            $model->switchId = $map['SwitchId'];
+        }
+        if (isset($map['VpcId'])) {
+            $model->vpcId = $map['VpcId'];
+        }
+
+        return $model;
+    }
+}