+++ pai-dlc-20201203/src/Paidlc.php
@@ -53,7 +53,6 @@ use AlibabaCloud\SDK\Paidlc\V20201203\Models\UpdateTensorboardRequest;
 use AlibabaCloud\SDK\Paidlc\V20201203\Models\UpdateTensorboardResponse;
 use AlibabaCloud\Tea\Utils\Utils;
 use AlibabaCloud\Tea\Utils\Utils\RuntimeOptions;
-use Darabonba\GatewayPop\Client;
 use Darabonba\OpenApi\Models\OpenApiRequest;
 use Darabonba\OpenApi\Models\Params;
 use Darabonba\OpenApi\OpenApiClient;
@@ -63,9 +62,6 @@ class Paidlc extends OpenApiClient
     public function __construct($config)
     {
         parent::__construct($config);
-        $this->_productId    = 'pai-dlc';
-        $gatewayClient       = new Client();
-        $this->_spi          = $gatewayClient;
         $this->_endpointRule = 'regional';
         $this->_endpointMap  = [
             'ap-northeast-1'              => 'pai-dlc.aliyuncs.com',
@@ -236,11 +232,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return CreateJobResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return CreateJobResponse::fromMap($this->execute($params, $req, $runtime));
+        return CreateJobResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -346,11 +339,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return CreateTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return CreateTensorboardResponse::fromMap($this->execute($params, $req, $runtime));
+        return CreateTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -393,11 +383,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return DeleteJobResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return DeleteJobResponse::fromMap($this->execute($params, $req, $runtime));
+        return DeleteJobResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -447,11 +434,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return DeleteTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return DeleteTensorboardResponse::fromMap($this->execute($params, $req, $runtime));
+        return DeleteTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -502,11 +486,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetJobResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetJobResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetJobResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -563,11 +544,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetJobEventsResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetJobEventsResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetJobEventsResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -630,11 +608,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetJobMetricsResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetJobMetricsResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetJobMetricsResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -691,11 +666,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetJobSanityCheckResultResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetJobSanityCheckResultResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetJobSanityCheckResultResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -756,11 +728,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetPodEventsResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetPodEventsResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetPodEventsResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -825,11 +794,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetPodLogsResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetPodLogsResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetPodLogsResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -887,11 +853,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetTensorboardResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -942,11 +905,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetTensorboardSharedUrlResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetTensorboardSharedUrlResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetTensorboardSharedUrlResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1002,11 +962,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetTokenResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetTokenResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetTokenResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1060,11 +1017,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return GetWebTerminalResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return GetWebTerminalResponse::fromMap($this->execute($params, $req, $runtime));
+        return GetWebTerminalResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1133,11 +1087,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return ListEcsSpecsResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return ListEcsSpecsResponse::fromMap($this->execute($params, $req, $runtime));
+        return ListEcsSpecsResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1187,11 +1138,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return ListJobSanityCheckResultsResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return ListJobSanityCheckResultsResponse::fromMap($this->execute($params, $req, $runtime));
+        return ListJobSanityCheckResultsResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1264,6 +1212,9 @@ class Paidlc extends OpenApiClient
         if (!Utils::isUnset($request->pageSize)) {
             $query['PageSize'] = $request->pageSize;
         }
+        if (!Utils::isUnset($request->paymentType)) {
+            $query['PaymentType'] = $request->paymentType;
+        }
         if (!Utils::isUnset($request->pipelineId)) {
             $query['PipelineId'] = $request->pipelineId;
         }
@@ -1312,11 +1263,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return ListJobsResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return ListJobsResponse::fromMap($this->execute($params, $req, $runtime));
+        return ListJobsResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1422,11 +1370,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return ListTensorboardsResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return ListTensorboardsResponse::fromMap($this->execute($params, $req, $runtime));
+        return ListTensorboardsResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1476,11 +1421,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return StartTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return StartTensorboardResponse::fromMap($this->execute($params, $req, $runtime));
+        return StartTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1524,11 +1466,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return StopJobResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return StopJobResponse::fromMap($this->execute($params, $req, $runtime));
+        return StopJobResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1578,11 +1517,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return StopTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return StopTensorboardResponse::fromMap($this->execute($params, $req, $runtime));
+        return StopTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1636,11 +1572,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return UpdateJobResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return UpdateJobResponse::fromMap($this->execute($params, $req, $runtime));
+        return UpdateJobResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**
@@ -1700,11 +1633,8 @@ class Paidlc extends OpenApiClient
             'reqBodyType' => 'json',
             'bodyType'    => 'json',
         ]);
-        if (Utils::isUnset($this->_signatureVersion) || !Utils::equalString($this->_signatureVersion, 'v4')) {
-            return UpdateTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
-        }
 
-        return UpdateTensorboardResponse::fromMap($this->execute($params, $req, $runtime));
+        return UpdateTensorboardResponse::fromMap($this->callApi($params, $req, $runtime));
     }
 
     /**